# api-docs/swagger.yaml
openapi: 3.0.3
info:
  title: CCTV Smart Attendance System API
  description: AI-powered attendance management with facial recognition, PPE detection, salary/fine management, and HR analytics.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login
  parameters:
    pageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    limitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    startDateParam:
      name: startDate
      in: query
      schema: { type: string, format: date }
    endDateParam:
      name: endDate
      in: query
      schema: { type: string, format: date }
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        code: { type: string }
    PaginatedResponse:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    User:
      type: object
      properties:
        _id: { type: string }
        employeeId: { type: string, example: EMP001 }
        firstName: { type: string, example: John }
        lastName: { type: string, example: Doe }
        email: { type: string, format: email }
        phone: { type: string }
        department: { type: string }
        designation: { type: string }
        baseSalary: { type: number, minimum: 0 }
        isActive: { type: boolean }
        role: { type: string, enum: [employee, admin, super-admin] }
        createdAt: { type: string, format: date-time }
    Attendance:
      type: object
      properties:
        _id: { type: string }
        userId: { type: string }
        date: { type: string, format: date }
        status: { type: string, enum: [present, absent, late, half-day, early-departure, holiday, leave] }
        firstDetection: { type: string, format: date-time }
        lastDetection: { type: string, format: date-time }
        totalWorkHours: { type: number }
        breakHours: { type: number }
        effectiveWorkHours: { type: number }
        shiftCompliance:
          type: object
          properties:
            wasLate: { type: boolean }
            lateMinutes: { type: number }
            wasEarlyDeparture: { type: boolean }
            earlyDepartureMinutes: { type: number }
            overtimeHours: { type: number }
        detections:
          type: array
          items:
            type: object
            properties:
              timestamp: { type: string, format: date-time }
              cameraId: { type: string }
              confidence: { type: number }
              livenessScore: { type: number, minimum: 0, maximum: 1 }
              emotion: { type: string, enum: [neutral, happy, stressed, fatigued, angry] }
              ppeCompliance:
                type: object
                properties:
                  helmet: { type: boolean }
                  vest: { type: boolean }
                  goggles: { type: boolean }
                  gloves: { type: boolean }
...
    Camera:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        rtspUrl: { type: string }
        status: { type: string, enum: [online, offline, maintenance] }
        location:
          type: object
          properties:
            zone: { type: string }
            floor: { type: number }
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  refreshToken: { type: string }
                  expiresIn: { type: string, example: "24h" }
                  user: { $ref: '#/components/schemas/User' }
        400: { $ref: '#/components/responses/BadRequest' }
        401: { $ref: '#/components/responses/Unauthorized' }
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employeeId, firstName, lastName, email, password, department]
              properties:
                employeeId: { type: string }
                firstName: { type: string }
                lastName: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                department: { type: string }
                role: { type: string, enum: [employee, admin, super-admin], default: employee }
      responses:
        201:
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        409: { description: Duplicate email or employee ID }
  /auth/refresh-token:
    post:
      tags: [Auth]
      summary: Refresh an expired access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        200:
          description: New token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  expiresIn: { type: string }
        401: { $ref: '#/components/responses/Unauthorized' }
  /auth/profile:
    get:
      tags: [Auth]
      summary: Get current user profile
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
    put:
      tags: [Auth]
      summary: Update current user profile (firstName, lastName, phone only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName: { type: string }
                lastName: { type: string }
                phone: { type: string }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }

  # ─── Employees ─────────────────────────────────────────────────
...
  /attendance:
    get:
      tags: [Attendance]
      summary: List attendance records (with filters)
      parameters:
        - { $ref: '#/components/parameters/pageParam' }
        - { $ref: '#/components/parameters/limitParam' }
        - { $ref: '#/components/parameters/startDateParam' }
        - { $ref: '#/components/parameters/endDateParam' }
        - { name: status, in: query, schema: { type: string, enum: [present, absent, late, half-day] } }
      responses:
        200:
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/PaginatedResponse' }
                  - type: object
                    properties:
                      records: { type: array, items: { $ref: '#/components/schemas/Attendance' } }
  /attendance/today/{userId}:
    get:
      tags: [Attendance]
      summary: Get today's attendance for a user
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  attendance: { $ref: '#/components/schemas/Attendance' }
  /attendance/report/{userId}:
    get:
      tags: [Attendance]
      summary: Get attendance report for a user
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
        - { $ref: '#/components/parameters/startDateParam' }
        - { $ref: '#/components/parameters/endDateParam' }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  attendances: { type: array, items: { $ref: '#/components/schemas/Attendance' } }
                  summary:
                    type: object
                    properties:
                      totalDays: { type: integer }
                      present: { type: integer }
                      late: { type: integer }
                      absent: { type: integer }
                      halfDay: { type: integer }
                      totalWorkHours: { type: number }
                      averageWorkHours: { type: number }
  /attendance/verify-face:
    post:
      tags: [Attendance]
      summary: Verify face and record attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [imageBuffer]
              properties:
                imageBuffer: { type: string, description: Base64 image }
                cameraId: { type: string }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  recognized: { type: boolean }
                  results: { type: array, items: { type: object } }
                  attendanceResults: { type: array, items: { type: object } }

  # ─── Cameras ───────────────────────────────────────────────────
  /cameras:
    get:
      tags: [Cameras]
      summary: List cameras
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [online, offline, maintenance] } }
        - { name: zone, in: query, schema: { type: string } }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  cameras: { type: array, items: { $ref: '#/components/schemas/Camera' } }
    post:
      tags: [Cameras]
      summary: Add camera (admin only)
      responses:
        201:
          description: Camera created
  /cameras/{id}/connect:
    post:
      tags: [Cameras]
      summary: Connect camera stream (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        200: { description: Camera connected }
  /cameras/{id}/disconnect:
    post:
      tags: [Cameras]
      summary: Disconnect camera stream (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        200: { description: Camera disconnected }

  # ─── Dashboard ─────────────────────────────────────────────────
...
  /analytics/predict/{date}:
    get:
      tags: [Analytics]
      summary: Predict attendance for a date
      parameters:
        - { name: date, in: path, required: true, schema: { type: string, format: date } }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  prediction: { type: object }
  /analytics/anomalies/{userId}/{date}:
    get:
      tags: [Analytics]
      summary: Detect anomalies for user on date
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
        - { name: date, in: path, required: true, schema: { type: string, format: date } }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  anomalies: { type: array, items: { type: object } }
...